release_notes_to_udm/cloud_function/requirements.txt
functions-framework==3.*
requests == 2.26.0
google-auth==2.6.0
google-cloud-secret-manager==2.10.0
xmltodict==0.13.0
google-api-python-client==2.65.0
google-cloud-logging==3.5.0
